# Implement multiple suffixes extension support with tests (#214)