--- a/src/internal.rs
+++ b/src/internal.rs
@@ -16,6 +16,7 @@ use exec::CommandTemplate;
 use lscolors::LsColors;
 use walk::FileType;
 use regex_syntax::{Expr, ExprBuilder};
+use regex::RegexSet;
 
 /// Configuration options for *fd*.
 pub struct FdOptions {
@@ -65,7 +66,7 @@ pub struct FdOptions {
     /// The extension to search for. Only entries matching the extension will be included.
     ///
     /// The value (if present) will be a lowercase string without leading dots.
-    pub extensions: Option<HashSet<String>>,
+    pub extensions: Option<RegexSet>,
 
     /// If a value is supplied, each item found will be used to generate and execute commands.
     pub command: Option<CommandTemplate>,