--- a/src/main.rs
+++ b/src/main.rs
@@ -25,6 +25,7 @@ mod app;
 mod exec;
 mod internal;
 mod output;
+mod utils;
 mod walk;
 
 #[cfg(windows)]
@@ -152,7 +153,7 @@ fn main() {
                 .collect(),
         },
         extensions: matches.values_of("extension").map(|exts| {
-            exts.map(|e| e.trim_left_matches('.').to_lowercase())
+            exts.map(|e| String::from(".") + &e.trim_left_matches('.'))
                 .collect()
         }),
         command,