Instead of relying on `Path.extension()`, the path is checked to see if it ends with any of the given extensions. The only quirk was to compare `.ext` to the end of the path instead of `ext`, or else we could have matched wrong files.

I had several ways of doing so in mind, and while not being the *quicket* this one should be the most obvious, and I don't think it's really necessary to micro-optimize here.

I also added a few test cases that should prevent regressions.