@sharkdp 2018-01-11
Welcome and thank you very much for your contribution!

> I had several ways of doing so in mind, and while not being the quicket this one should be the most obvious, and I don't think it's really necessary to micro-optimize here.

Unfortunately, I think it will be. I have been running some benchmarks and it looks like matching on file extensions is about a factor of two slower with this PR:

Before:
```
benchmarking bench/fd -HI -e jpg '' '/home/shark'
time                 834.3 ms   (737.0 ms .. 878.1 ms, ci 0.990)
                     0.998 R²   (0.996 R² .. 1.000 R², ci 0.990)
mean                 861.9 ms   (839.3 ms .. 873.2 ms, ci 0.990)
std dev              19.55 ms   (0.0 s .. 19.68 ms, ci 0.990)
variance introduced by outliers: 19% (moderately inflated)
```

After (this PR):
```
benchmarking bench/fd -HI -e jpg '' '/home/shark'
time                 1.777 s    (1.704 s .. 1.840 s, ci 0.990)
                     1.000 R²   (1.000 R² .. 1.000 R², ci 0.990)
mean                 1.757 s    (1.740 s .. 1.776 s, ci 0.990)
std dev              17.99 ms   (0.0 s .. 20.65 ms, ci 0.990)
variance introduced by outliers: 19% (moderately inflated)
```