@althonos 2018-01-11
@sharkdp : I just thought of something that should work better. Don't merge yet.