@sharkdp 2018-01-12
Thanks!

Unfortunately, this is still much slower than the original version:
```
benchmarking bench/fd -HI -e jpg '' '/home/shark'
time                 1.681 s    (1.667 s .. 1.701 s, ci 0.990)
                     1.000 R²   (1.000 R² .. 1.000 R², ci 0.990)
mean                 1.683 s    (1.679 s .. 1.685 s, ci 0.990)
std dev              3.412 ms   (0.0 s .. 3.603 ms, ci 0.990)
variance introduced by outliers: 19% (moderately inflated)
```

It seems like the expensive part is the `.to_string_lossy().to_lowercase()` conversion that needs to be done for each and every path(!). I found a way to make it faster by only selecting the file name first:
``` rust
if let Some(os_path) = entry_path
    .file_name()
    .map(|s| s.to_string_lossy().to_lowercase())
{
    if !filter_exts.iter().any(|x| os_path.ends_with(x)) {
        return ignore::WalkState::Continue;
    }
}
```
.. but it's still not ideal:
```
benchmarking bench/fd -HI -e jpg '' '/home/shark'
time                 926.8 ms   (NaN s .. 945.7 ms, ci 0.990)
                     1.000 R²   (1.000 R² .. 1.000 R², ci 0.990)
mean                 934.5 ms   (932.4 ms .. 936.8 ms, ci 0.990)
std dev              2.233 ms   (0.0 s .. 2.568 ms, ci 0.990)
variance introduced by outliers: 19% (moderately inflated)
```

Any more ideas? :smile: