@althonos 2018-01-12
@sharkdp : I may have an idea that avoids using `to_string_lossy` and uses only `extension`...