@sharkdp 2018-01-17
Wow, impressive work! :fireworks: 

This is indeed faster than before. My only concern is that this increases the code complexity quite a bit :smile:.

I have added a few inline comments.

Also, sorry for the late reply. I have been working on a way to perform benchmarks for `fd` in an easier way. I wrote a new tool - [hyperfine](https://github.com/sharkdp/hyperfine) - that can perform benchmarks with disk-cache-warmups (or diskcache-reset commands between the benchmarks). If anyone is interested, I've uploaded my benchmarking scripts for `fd` here: https://github.com/sharkdp/fd-benchmarks