@althonos 2018-01-17
> Wow, impressive work! 🎆

I'm the one who should be saying that ! I'm messing up a bit with a PR, you end writing a complete tool from scratch 😅 

> My only concern is that this increases the code complexity quite a bit 😄.

I agree, the Windows part in particular was not so easy because of [`EncodeWide`](https://doc.rust-lang.org/std/os/windows/ffi/struct.EncodeWide.html) not being a `DoubleEndedIterator` (which it should be IMO). So maybe this could be reported to the Rust team.

In the end, this is just a dirty way of implementing `Into<&OsStr>.ends_with(Into<&OsStr>)` (now that I'm writing it, I realize it could be a lot cleaner to do that with a custom trait).