@sharkdp 2018-01-24
Fantastic, thank you very much.

I'm just waiting for the Mac builds to finish (which might take some - there seems to be a maintenance downtime for Mac builds).