@sharkdp 2018-01-24
Looks like the new Unicode test actually fails on macOS.