@sharkdp 2018-01-24
Also, I just noticed that there is a change in behavior concerning hidden files / dotfiles:

Before:
```  bash
> touch .hidden
> fd -He hidden
# no match
```

After:
```
> touch .hidden
> fd -He hidden
.hidden
```