@sharkdp 2018-01-24
I think the Unicode test failing on macOS could be due to https://stackoverflow.com/a/6153713/704831