@althonos 2018-01-25
@sharkdp : in the end, I think we could simply:
* use the current implementation on Linux (+ non MacOS UNIXes ?)
* use a fallback with `to_string_lossy` on the filename for MacOS + Windows (I don't think the reverse UTF-16 iterator is that efficient actually)

Maybe we could also use [`to_str`](https://doc.rust-lang.org/std/ffi/struct.OsStr.html#method.to_str) when possible, to avoid doing a copy with valid Unicode path names.