@sharkdp 2018-01-26
@althonos Thank you for the updates. It still seems to fail on macOS, though.