@althonos 2018-01-29
I'm having trouble even understanding what's the issue in the MacOS build on Travis. Since I don't have a Mac to code on, this is getting more and more complicated for me to test.

At the beginning this PR was just supposed to bring multiple extensions comparison to `fd`, but now I end up porting the whole code to support UTF-8, even though I don't think it even works as intended for the program on MacOS in its current state.