@sharkdp 2018-01-29
> At the beginning this PR was just supposed to bring multiple extensions comparison to fd, but now I end up porting the whole code to support UTF-8, even though I don't think it even works as intended for the program on MacOS in its current state.

I'm sorry - you are right (I have not seen your "Edit" comment above until now). The normalization issue seems to be a general problem on macOS. 

I just wanted to have a Unicode test because I was worried that the specialized iterators would do something wrong with Unicode characters.

I suggest the following for now:
- We replace replace `é` by `é` (the normalized version) in the tests for now. This should work on macOS as well.
- I will create a new ticket to solve the normalization issue on macOS in general.

In general, I'm still not completely happy with the split-up version... couldn't we write something like this for all operating systems:
``` rust
/// Determine if an os string ends with any of the given extensions (case insensitive).
pub fn path_has_any_extension<'a, I>(path: &Path, exts: I) -> bool
where
    I: 'a + IntoIterator<Item = &'a String>,
{
    // TODO: remove these two lines when we drop support for Rust version < 1.23.
    #[allow(unused_imports)]
    use std::ascii::AsciiExt;

    if let Some(path_str) = path.to_str() {
        exts.into_iter().any(|x| {
            let mut it = path_str.chars().rev();

            if x.chars()
                .rev()
                .zip(&mut it)
                .all(|(a, b)| a.eq_ignore_ascii_case(&b))
            {
                match it.next() {
                    Some('/') | None => false,
                    _ => true,
                }
            } else {
                false
            }
        })
    } else {
        false
    }
}
```
(yes, this will not work if a filename has invalid unicode characters)