@althonos 2018-01-29
No need to be sorry, I didn't mean to be rude ! I think what you came up with should work as well, so I'll edit the tests to reflect that. 

> In general, I'm still not completely happy with the split-up version...

The advantage of the split-up version is that it's zero-copy on UNIXes, so I think it will end up being quicker. As a side-note, the current implementation could maybe work for both Linux and MacOS if instead of normalising the paths, we normalized the extensions in a way HFS+ likes beforehand. So only Windows would end up using the default slower way.